improved instance freezing reset speed (scuffed)
added slightly scuffed system that unfreezes instances when you go to wall, and doesn't freeze instances while on the wall. This improves reset speed SIGNIFICANTLY, at the cost of significantly increased memory usage. Instances also do not refreeze when you do a normal focus onto an instance, but if you lock an instance, it will freeze.